// src/foxess/lib/api/v1.d.ts
export interface paths {
// ...
  "/op/v0/device/list": {
    post: {
      parameters: {
        header: {
          "Content-Type": string;
          /** @description Generate apikey from the API management function of the platform */
          token: string;
          /** @description Signature rule: Encrypt the string url + \"\r\n\" + token + \"\r\n\" + timestamp with md5 */
          signature: string;
          /** @description Current timestamp */
          timestamp: string;
          /**
           * @description Language
           * @example en
           */
          lang: string;
        };
      };
      requestBody: {
        content: {
          "application/json": {
            /**
             * @description Current page number
             * @default 1
             */
            currentPage: number;
            /**
             * @description The amount of data presented per page
             * @default 10
             */
            pageSize: number;
          };
        };
      };
      responses: {
        /** @description OK */
        200: {
          content: {
            "application/json": {
              /** @description Error number (When the result is not equal to zero, the request fails) */
              errno: number;
              result: {
                /** @description Current page number */
                currentPage: number;
                /** @description The amount of data presented per page */
                pageSize: number;
                /** @description Data volume total */
                total: number;
                data: {
                    /** @description Serial number of the inverter */
                    deviceSN: string;
                    /** @description Serial number of the Data Logger */
                    moduleSN: string;
                    /** @description Id of power station */
                    plantID: string;
                    /** @description Status of module */
                    status: number;
                    /** @description Is there a photovoltaic system available */
                    hasPV: boolean;
                    /** @description Is there a battery available */
                    hasBattery: boolean;
                    /** @description Type of device */
                    deviceType: string;
                    /** @description Type of product */
                    productType: string;
                  }[];
              };
            };
          };
        };
      };
    };
    parameters: {
      header: {
        "Content-Type": string;
        /** @description Generate apikey from the API management function of the platform */
        token: string;
        /** @description Signature rule: Encrypt the string url + \"\r\n\" + token + \"\r\n\" + timestamp with md5 */
        signature: string;
        /** @description Current timestamp */
        timestamp: string;
        /**
         * @description Language
         * @example en
         */
        lang: string;
      };
    };
  };
// ...
  "/op/v0/device/real/query": {
    post: {
      parameters: {
        header: {
          "Content-Type": string;
          /** @description Generate apikey from the API management function of the platform */
          token: string;
          /** @description Signature rule: Encrypt the string url + \"\r\n\" + token + \"\r\n\" + timestamp with md5 */
          signature: string;
          /** @description Current timestamp */
          timestamp: string;
          /**
           * @description Language
           * @example en
           */
          lang: string;
        };
      };
      requestBody: {
        content: {
          "application/json": {
            /** @description If this parameter is not passed, all variable data is obtained by default item type: string */
            variables?: string[];
            /** @description Serial Number of Inverter If no parameter is specified, all devices are obtained */
            sn?: string;
          };
        };
      };
      responses: {
        /** @description OK */
        200: {
          content: {
            "application/json": {
              errno: number;
              result: {
                  /** @description Serial Number of Inverter */
                  deviceSN: string;
                  datas: {
                      /** @description Variable name If the data is not found, it will not be returned */
                      variable: string;
                      /** @description Unit */
                      unit: string;
                      /** @description Name in English */
                      name: string;
                      /** @description Value */
                      value: number;
                    }[];
                  /** @description Time of Data Update, utc time */
                  time: string;
                }[];
            };
          };
        };
      };
    };
    parameters: {
      header: {
        "Content-Type": string;
        /** @description Generate apikey from the API management function of the platform */
        token: string;
        /** @description Signature rule: Encrypt the string url + \"\r\n\" + token + \"\r\n\" + timestamp with md5 */
        signature: string;
        /** @description Current timestamp */
        timestamp: string;
        /**
         * @description Language
         * @example en
         */
        lang: string;
      };
    };
  };
// ...
}